Optimize the way we handle ctx.read() in the read loop when AUTO_READ is false in the native transports
Motivation:

We currently only continue reading when AUTO_READ is true and not correctly respect the ctx.read() that may be done during the read-loop. In this case we can just continue reading and not need to go back to the kqueue / epoll and pay for the wakeup.

Modifications:

- Adjust code to correctly take readPending into account by moving it to the RecvHandle implementations of the native transports.
- Adjust testcase to correctly take into account that multiple reads may happen during one read loop.

Result:

Less overhead if user manages reads directly and set AUTO_READ to false.